Development branch

+++ Modules/DispatchOperation/app/Http/Controllers/DropController.php
@@ -3,8 +3,10 @@
 namespace Modules\DispatchOperation\Http\Controllers;
 
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 use Modules\DispatchOperation\Classes\Data\Request\CreateDropData;
 use Modules\DispatchOperation\Classes\Data\Request\EditDropData;
+use Modules\DispatchOperation\Models\Dispatch;
 use Modules\DispatchOperation\Services\DropService;
 
 class DropController extends Controller
@@ -15,6 +17,8 @@ class DropController extends Controller
 
     public function store(CreateDropData $data)
     {
+        Gate::authorize('create', Dispatch::class);
+
         $this->dropService->createDrop($data);
 
         return back()->with('success', 'Drop added successfully.');
@@ -22,6 +26,8 @@ class DropController extends Controller
 
     public function update(EditDropData $data, int $id)
     {
+        Gate::authorize('update', Dispatch::class);
+
         $this->dropService->updateDrop($data, $id);
 
         return back()->with('success', 'Drop updated successfully.');
@@ -29,6 +35,8 @@ class DropController extends Controller
 
     public function destroy(int $id)
     {
+        Gate::authorize('delete', Dispatch::class);
+
         $this->dropService->deleteDrop($id);
 
         return back()->with('success', 'Drop deleted successfully.');
